pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "m1-cis"
version = "0.2.0"
authors = [
  { name="Michał Kozłowski", email="m1chol@proton.me" },
]
description = "Context Image Search Utility"
dependencies = [
  "google-genai>=1.38.0",
  "transformers>=4.56.1",
  "torch>=2.8.0",
  "torchvision>=0.23.0",
  "pydantic>=2.11.9",
  "Pillow>=11.3.0",
  "requests>=2.32.5"
]
readme = "README_PYPI.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search"

]
license = "Apache-2.0"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/M1chol/m1-cis"
Issues = "https://github.com/M1chol/m1-cis/issues"